ex7.9 Refactor code
This refactoring has highlighted the need of a separate module that satisfies a Web API.
This module which I will call webmusic can be used to develop
a frontend that calls the implementation of an HTTP handler
for sorting requests.
The code will be very similar except that the test will not
concern itself with music tracks but with a webmusic object.